+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # main
 
+* Added cop `Rails/DestroyAllBang` ([#X](https://github.com/petalmd/rubocop-petal/pull/X))
 * Fix `Migration/ChangeTableReferences` offense location (Fix [#61](https://github.com/petalmd/rubocop-petal/issues/61)) ([#62](https://github.com/petalmd/rubocop-petal/pull/62))
 * Added cop `Sidekiq/NoEarlyNilReturn` ([#58](https://github.com/petalmd/rubocop-petal/pull/58))
 * Added cop `Rails/EnumStartingValue` ([#57](https://github.com/petalmd/rubocop-petal/pull/57))